requirements.txt
streamlit
opencv-python
ultralytics
Pillow
numpy
requests
